Adding @wasade's parallel to per sample fasta (#4)
* Adding @wasade's parallel to per sample fasta

* Fixing docs

* Adding FASTQ output to per sample files and reduce code duplication

* fq -> fastq

* Trying to fix coveralls

+++ qiita_files/tests/test_demux.py
@@ -11,6 +11,8 @@ from __future__ import division
 import os
 import tempfile
 from unittest import TestCase, main
+from functools import partial
+from shutil import rmtree
 
 import h5py
 import numpy as np
@@ -19,7 +21,8 @@ import numpy.testing as npt
 from qiita_files.demux import (buffer1d, buffer2d, _has_qual,
                                _per_sample_lengths, _summarize_lengths,
                                _set_attr_stats, _construct_datasets, to_hdf5,
-                               to_ascii, stat, to_per_sample_ascii)
+                               to_ascii, stat, to_per_sample_ascii,
+                               to_per_sample_files)
 
 
 class BufferTests(TestCase):
@@ -126,7 +129,11 @@ class DemuxTests(TestCase):
     def tearDown(self):
         self.hdf5_file.close()
         for f in self.to_remove:
-            os.remove(f)
+            if os.path.exists(f):
+                if os.path.isdir(f):
+                    rmtree(f)
+                else:
+                    os.remove(f)
 
     def test_has_qual(self):
         with tempfile.NamedTemporaryFile('r+', suffix='.fna') as f:
@@ -320,6 +327,69 @@ class DemuxTests(TestCase):
 
         obs = [(s[0], list(s[1])) for s in to_per_sample_ascii(self.hdf5_file)]
         self.assertEqual(obs, exp)
+
+    def test_to_files(self):
+        # implicitly tested with test_to_per_sample_fasta
+        pass
+
+    def test_to_per_sample_files(self):
+        with tempfile.NamedTemporaryFile('r+', suffix='.fq',
+                                         delete=False) as f:
+            f.write(fqdata_variable_length)
+
+        self.to_remove.append(f.name)
+
+        with tempfile.NamedTemporaryFile('r+', suffix='.demux',
+                                         delete=False) as demux_f:
+            pass
+
+        self.to_remove.append(demux_f.name)
+
+        with h5py.File(demux_f.name, 'w') as demux:
+            to_hdf5(f.name, demux)
+
+        tmp_dir = tempfile.mkdtemp()
+        self.to_remove.append(tmp_dir)
+        path_builder = partial(os.path.join, tmp_dir)
+
+        # Test to fastq
+        to_per_sample_files(demux_f.name, out_dir=tmp_dir, n_jobs=1,
+                            out_format='fastq')
+        sample_a_path = path_builder("a.fastq")
+        sample_b_path = path_builder("b.fastq")
+        self.assertTrue(os.path.exists(sample_a_path))
+        self.assertTrue(os.path.exists(sample_b_path))
+
+        with open(sample_a_path, 'rb') as af:
+            obs = af.read()
+        self.assertEqual(
+            obs, b'@a_0 orig_bc=abc new_bc=abc bc_diffs=0\nxyz\n+\nABC\n')
+
+        with open(sample_b_path, 'rb') as bf:
+            obs = bf.read()
+        self.assertEqual(
+            obs, b'@b_0 orig_bc=abw new_bc=wbc bc_diffs=4\nqwe\n+\nDFG\n'
+                 b'@b_1 orig_bc=abw new_bc=wbc bc_diffs=4\nqwexx\n+\nDEF#G\n')
+
+        # Test to fasta and parallel
+        to_per_sample_files(demux_f.name, out_dir=tmp_dir, n_jobs=2,
+                            out_format='fasta')
+
+        sample_a_path = path_builder("a.fna")
+        sample_b_path = path_builder("b.fna")
+        self.assertTrue(os.path.exists(sample_a_path))
+        self.assertTrue(os.path.exists(sample_b_path))
+
+        with open(sample_a_path, 'rb') as af:
+            obs = af.read()
+        self.assertEqual(
+            obs, b'>a_0 orig_bc=abc new_bc=abc bc_diffs=0\nxyz\n')
+
+        with open(sample_b_path, 'rb') as bf:
+            obs = bf.read()
+        self.assertEqual(
+            obs, b'>b_0 orig_bc=abw new_bc=wbc bc_diffs=4\nqwe\n'
+                 b'>b_1 orig_bc=abw new_bc=wbc bc_diffs=4\nqwexx\n')
 
     def test_fetch(self):
         # implicitly tested with test_to_ascii
